+++ lib/LinearFMWaveform.cc
@@ -3,19 +3,19 @@
 /*
  * Generate a PRI of data for the given waveform
  */
-std::vector<std::complex<float>> LinearFMWaveform::sample() {
+std::vector<std::complex<double>> LinearFMWaveform::sample() {
   // Sample interval
-  float ts = 1 / sampRate;
+  double ts = 1 / sampRate;
   // Number of samples per pulse
   int nSampsPulse = static_cast<int>(sampRate * pulsewidth);
   int nSampsPri = static_cast<int>(sampRate / prf);
-  float t;
-  std::vector<std::complex<float>> wave(nSampsPri, 0);
+  double t;
+  std::vector<std::complex<double>> wave(nSampsPri, 0);
   for (int n = 0; n < nSampsPulse; n++) {
     t = n * ts;
-    float phase =
+    double phase =
         -bandwidth / 2 * t + bandwidth / (2 * pulsewidth) * std::pow(t, 2);
-    wave[n] = std::exp(Im * (float)(2 * M_PI) * phase);
+    wave[n] = std::exp(Im * (double)(2 * M_PI) * phase);
   }
   return wave;
 }
